+++ astroquery/gemini/core.py
@@ -8,22 +8,17 @@ import os
 
 from datetime import date
 
-from astropy import log
+from astroquery import log
 from astropy import units
 from astropy.table import Table, MaskedColumn
 
 from astroquery.gemini.urlhelper import URLHelper
 import numpy as np
 
-import logging
-
 from ..query import BaseQuery, QueryWithLogin
 from ..utils.class_or_instance import class_or_instance
 from . import conf
 from ..exceptions import AuthenticationWarning
-
-
-logger = logging.getLogger(__name__)
 
 
 __all__ = ['Observations', 'ObservationsClass']  # specifies what to import
@@ -133,7 +128,7 @@ class ObservationsClass(QueryWithLogin):
         params = dict(username=username, password=password)
         r = self._session.request('POST', 'https://archive.gemini.edu/login/', params=params)
         if b'<P>Welcome, you are sucessfully logged in' not in r.content:
-            logger.error('Unable to login, please check your credentials')
+            log.error('Unable to login, please check your credentials')
             return False
         return True